add SherpaDocHelper
for more easily documenting types used in your functions.

this still is not perfect. we should be using javadoc and other
comments around types, and include that in de documentation
automatically.  also, you now need to list the types explicitly
yourself. the exported types should be found automatically by
sherpadoc.